nmap_api/requirements.txt
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-nmap==0.7.1
Werkzeug==2.2.3
